Move MainTitle to admin's components
