+++ src/pages/Admin/Home/Dashboard/TaskDetail/Company/CompanyDetailInfo/MainTitle/container.tsx
@@ -3,7 +3,7 @@ import React, { FunctionComponent } from "react";
 import { ICompany } from "$interfaces/Company";
 import { IUser } from "$interfaces/User";
 
-import { MainTitle } from "../../../MainTitle";
+import { MainTitle } from "../../../../../../components/MainTitle";
 import { useTranslations } from "$hooks/queries";
 
 export const MainTitleContainer: FunctionComponent<IMainTitleContainerProps> = ({ company }) => {
