+++ src/pages/Admin/Home/Dashboard/TaskDetail/Offer/OfferDetailInfo/MainTitle/container.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from "react";
 
-import { MainTitle } from "../../../MainTitle";
+import { MainTitle } from "../../../../../../components/MainTitle";
 import { useTranslations } from "$hooks/queries";
 import { IOffer } from "$interfaces/Offer";
